+++ .sim/run.sh
@@ -1,5 +1,0 @@
-reps-sim-01.sh
-reps-sim-02.sh
-reps-sim-03.sh
-reps-sim-04.sh
-reps-sim-05.sh